Refactor tests for DefaultIfNotSingle in ImmutableArray
Removed the test for multiple matches from ImmutableArrayExtensionsTests.cs. Added new tests in EnumerableExtensionsTests.cs to cover scenarios for empty arrays, null predicates, and single matches. Updated the overall structure to focus on edge cases for the DefaultIfNotSingle method.
