+++ tests/Moq.Analyzers.Test/Common/EnumerableExtensionsTests.cs
@@ -3,19 +3,71 @@ namespace Moq.Analyzers.Test.Common;
 public class EnumerableExtensionsTests
 {
     [Fact]
-    public void DefaultIfNotSingle_ReturnsNull_WhenSourceIsNull()
+    public void DefaultIfNotSingle_IEnumerable_ThrowsArgumentNullException_WhenPredicateIsNull()
     {
-        IEnumerable<object> source = null!;
-        object? result = source.DefaultIfNotSingle();
-        Assert.Null(result);
+        IEnumerable<string> source = ["a", "b", "c"];
+        ArgumentNullException ex = Assert.Throws<ArgumentNullException>(() => source.DefaultIfNotSingle(null!));
+        Assert.Equal("predicate", ex.ParamName);
     }
 
     [Fact]
-    public void DefaultIfNotSingle_ReturnsNull_WhenSourceIsEmpty()
+    public void DefaultIfNotSingle_ImmutableArray_CallsEnumerableExtension()
     {
-        IEnumerable<object> source = [];
-        object? result = source.DefaultIfNotSingle();
-        Assert.Null(result);
+        ImmutableArray<string> source = ImmutableArray.Create("a", "b", "c");
+        string? result = source.DefaultIfNotSingle(x => string.Equals(x, "b"));
+        Assert.Equal("b", result);
+    }
+
+    [Fact]
+    public void DefaultIfNotSingle_ImmutableArray_ReturnsElement_WhenSingleMatch()
+    {
+        ImmutableArray<int> source = [1, 2, 3];
+        int? result = source.DefaultIfNotSingle(x => x == 2);
+        Assert.Equal(2, result);
+    }
+
+    [Fact]
+    public void DefaultIfNotSingle_ImmutableArray_ReturnsNull_WhenEmpty()
+    {
+        ImmutableArray<int> source = ImmutableArray<int>.Empty;
+        int? result = source.DefaultIfNotSingle(x => x > 0);
+        Assert.Equal(default(int), result);
+    }
+
+    [Fact]
+    public void DefaultIfNotSingle_ImmutableArray_ReturnsNull_WhenIsDefaultOrEmpty()
+    {
+        ImmutableArray<int> source = ImmutableArray<int>.Empty;
+        int? result = source.DefaultIfNotSingle(x => true);
+        Assert.Equal(0, result);
+
+        ImmutableArray<int> defaultSource = default;
+        result = defaultSource.DefaultIfNotSingle(x => true);
+        Assert.Equal(0, result);
+    }
+
+    [Fact]
+    public void DefaultIfNotSingle_ImmutableArray_ReturnsNull_WhenMultipleMatches()
+    {
+        ImmutableArray<int> source = [1, 2, 2, 3];
+        int? result = source.DefaultIfNotSingle(x => x == 2);
+        Assert.Equal(0, result);
+    }
+
+    [Fact]
+    public void DefaultIfNotSingle_ImmutableArray_String_ThrowsArgumentNullException_WhenPredicateIsNull()
+    {
+        ImmutableArray<string> source = ImmutableArray.Create("a", "b", "c");
+        ArgumentNullException ex = Assert.Throws<ArgumentNullException>(() => source.DefaultIfNotSingle(null!));
+        Assert.Equal("predicate", ex.ParamName);
+    }
+
+    [Fact]
+    public void DefaultIfNotSingle_ImmutableArray_ThrowsArgumentNullException_WhenPredicateIsNull()
+    {
+        ImmutableArray<string> source = ["a", "b", "c"];
+        ArgumentNullException ex = Assert.Throws<ArgumentNullException>(() => source.DefaultIfNotSingle(null!));
+        Assert.Equal("predicate", ex.ParamName);
     }
 
     [Fact]
@@ -35,46 +87,18 @@ public class EnumerableExtensionsTests
     }
 
     [Fact]
-    public void DefaultIfNotSingle_IEnumerable_ThrowsArgumentNullException_WhenPredicateIsNull()
+    public void DefaultIfNotSingle_ReturnsNull_WhenSourceIsEmpty()
     {
-        IEnumerable<string> source = ["a", "b", "c"];
-        ArgumentNullException ex = Assert.Throws<ArgumentNullException>(() => source.DefaultIfNotSingle(null!));
-        Assert.Equal("predicate", ex.ParamName);
+        IEnumerable<object> source = [];
+        object? result = source.DefaultIfNotSingle();
+        Assert.Null(result);
     }
 
     [Fact]
-    public void DefaultIfNotSingle_ImmutableArray_ThrowsArgumentNullException_WhenPredicateIsNull()
+    public void DefaultIfNotSingle_ReturnsNull_WhenSourceIsNull()
     {
-        ImmutableArray<string> source = ["a", "b", "c"];
-        ArgumentNullException ex = Assert.Throws<ArgumentNullException>(() => source.DefaultIfNotSingle(null!));
-        Assert.Equal("predicate", ex.ParamName);
-    }
-
-    [Fact]
-    public void DefaultIfNotSingle_ImmutableArray_ReturnsNull_WhenIsDefaultOrEmpty()
-    {
-        ImmutableArray<int> source = ImmutableArray<int>.Empty;
-        int? result = source.DefaultIfNotSingle(x => true);
-        Assert.Equal(0, result);
-
-        ImmutableArray<int> defaultSource = default;
-        result = defaultSource.DefaultIfNotSingle(x => true);
-        Assert.Equal(0, result);
-    }
-
-    [Fact]
-    public void DefaultIfNotSingle_ImmutableArray_ReturnsElement_WhenSingleMatch()
-    {
-        ImmutableArray<int> source = [1, 2, 3];
-        int? result = source.DefaultIfNotSingle(x => x == 2);
-        Assert.Equal(2, result);
-    }
-
-    [Fact]
-    public void DefaultIfNotSingle_ImmutableArray_ReturnsNull_WhenMultipleMatches()
-    {
-        ImmutableArray<int> source = [1, 2, 2, 3];
-        int? result = source.DefaultIfNotSingle(x => x == 2);
-        Assert.Equal(0, result);
+        IEnumerable<object> source = null!;
+        object? result = source.DefaultIfNotSingle();
+        Assert.Null(result);
     }
 }
